requirements.txt
fastapi==0.105.0
uvicorn==0.24.0.post1
python-jose==3.3.0
bcrypt==4.0.1
passlib[bcrypt]==1.7.4
pymysql==1.1.0
cryptography==41.0.7
pydantic[email]==2.5.2
sqlalchemy==2.0.23
python-multipart==0.0.6
alembic==1.13.1

# Dependências de Teste
pytest==7.4.3
pytest-cov==4.1.0
httpx==0.25.2